Improvements to completion logic (mainly for top level) (#102)
Improvements to completion logic (mainly for top level). In true TDD
fashion, I'm fairly confident with the changes given the tests are all
passing.

Added additional tests for
- curly brace in default behavior
https://github.com/acao/codemirror-json-schema/issues/96
- autocomplete for a schema with top level $ref
- autocomplete for a schema with top level complex type

I wanted to also work on the "include value completions for boolean"
test case but there's a bit extra work (when completing the property
value and the node used for the completion is different from the
original node passed in to `getValueCompletions()`, we also need to
figure out the correct `from` and `to` values for the completion
context, so that applying the completion replaces the correct things,
otherwise it will replace the original node which may not be what we
want) to be done to get it working as expected, so decided to tackle
part of it but leave it commented out for now.