added limited support for conditional property values (#135)
As mentioned in #129 we don't get fully resolved schemas based on the
data. Following the discussion in
https://github.com/sagold/json-schema-library/issues/61, that appears to
be expected.

The workaround I implemented is to "fake" a deep schema resolution by
performing the shallow schema resolution on the direct children
properties of the schema (we only really care about the direct children
in this context).